Update contract formatting with prettier-plugin-solidity

+++ contracts/testing/TestGovernance.sol
@@ -2,13 +2,16 @@
 pragma solidity 0.8.3;
 
 import "../Governance.sol";
-contract TestGovernance is Governance{
-    event Received(address, uint);
+
+contract TestGovernance is Governance {
+    event Received(address, uint256);
+
     receive() external payable {
         emit Received(msg.sender, msg.value);
     }
+
     //allows the contract to recieve funds for gas via harhdat-impersonate account
-    fallback() external payable{
+    fallback() external payable {
         emit Received(msg.sender, msg.value);
     }
 }
